+++ config/app.js
@@ -73,7 +73,6 @@ app.use("/assets", express.static(base + "/node_modules/popper.js/dist/umd"))
 app.use("/assets", express.static(base + "/node_modules/font-awesome"))
 app.use("/assets", express.static(base + "/node_modules/bootstrap/dist"))
 app.use("/assets/js", express.static(base + "/node_modules/turbolinks/dist"))
-app.use("/assets/js", express.static(base + "/node_modules/d3/build"))
 app.use("/assets/js", express.static(base + "/app/assets/js"))
 
 /**