+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+
 # Change log
 
 All notable changes to this project will be documented in this file.
@@ -12,11 +13,12 @@ Changes are categorized as follows:
 
 ## [Unreleased]
 
-* 3 minor enhancements:
+* 4 minor enhancements:
 
   * docs: add github paths
   * deps: add debug
   * config: set json spaces 2
+  * middleware: nested parsers
 
 ## 2018-05-03, Version 1.0.1