+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Changes are categorized as follows:
 
 ## [Unreleased]
 
+## 2018-05-05, Version 1.1.0
+
 * 12 minor enhancements:
 
   * docs: add github paths